flux_hpc/requirements.txt
# Requirements для FLUX HPC генератора изображений

torch==2.7.1
torchvision>=0.18.0
diffusers>=0.30.0
transformers>=4.45.0
Pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.25.0

# Опциональные пакеты для оптимизации
xformers>=0.0.24  # Для оптимизации внимания (ускорение и снижение памяти)
safetensors>=0.4.0  # Для безопасной загрузки моделей
